pyproject.toml
[project]
name = "auto-scraper"
version = "0.1.0"
description = "Description"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
